+++ src/dataStructures/doubleLinkedList/utils/get.ts
@@ -1,5 +1,13 @@
 import { DoubleLinkedListNode } from '../doubleLinkedList';
-import { formatIndex } from './formatIndex';
+
+export function formatIndex(index: any, size: number): number {
+  const indexInt = index | 0;
+  if (indexInt < 0) {
+    return indexInt + size;
+  } else {
+    return indexInt;
+  }
+}
 
 export function get<T = unknown>(index: number, size: number, head: null | DoubleLinkedListNode<T>, tail: null | DoubleLinkedListNode<T>): null | DoubleLinkedListNode<T> {
   const position = formatIndex(index, size);