+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - **algorithm/array:** add removeElement ([6940656](https://github.com/air-supply94/js-algorithm/commit/69406564fc87b8918ecef93a4e2feea2677e6489))
 - **algorithm/array:** add twoSum ([a93da8f](https://github.com/air-supply94/js-algorithm/commit/a93da8fdd1a5cf9228a8cda70410b1e47433fc85))
 - **algorithm/backTrack:** add combine ([74ce4ed](https://github.com/air-supply94/js-algorithm/commit/74ce4eda583e9368d48e1e5948f371ada6064bc8))
+- **algorithm/backTrack:** add findLadders ([ca9ecec](https://github.com/air-supply94/js-algorithm/commit/ca9ecec1ac28ca1b603023f81305c96d02d6f3f3))
 - **algorithm/backTrack:** add solveSudoku ([a85601c](https://github.com/air-supply94/js-algorithm/commit/a85601cb2ed535d46e4efe08673a5188250a5f5e))
 - **algorithm/bfs:** add minDepth ([31ead4a](https://github.com/air-supply94/js-algorithm/commit/31ead4a45634cb14bc5ef9151856e93ee790f421))
 - **algorithm/bfs:** add openLock ([59d2a95](https://github.com/air-supply94/js-algorithm/commit/59d2a95acdbb5b2fb1b2f5eb8fceb9faa2582c64))