+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.0.0 (2021-07-30)
+# 0.0.0 (2021-08-04)
 
 ### Bug Fixes
 
@@ -119,6 +119,7 @@
 - **bst/leetcode:** add serializePostOrder ([86d2bba](https://github.com/air-supply94/js-algorithm/commit/86d2bba87b10dc35c6c216b3e0ec88f8ed2cf86d))
 - **bst/leetcode:** add serializePreAndInOrder ([3504aff](https://github.com/air-supply94/js-algorithm/commit/3504affb9f63de05912046066d8c12fe8bffe000))
 - **bst/leetcode:** update lowestCommonAncestor ([d2cecad](https://github.com/air-supply94/js-algorithm/commit/d2cecad5356c2e460b25fe30dfe1125346e20b6d))
+- **bst/leetcode:** 迁移到线上 leetcode ([1b7548f](https://github.com/air-supply94/js-algorithm/commit/1b7548f97661c3aaef76bd42bfd8cf133cae2dc5))
 - **bst/leetcode:** 迁移到线上 leetcode ([2bb8e77](https://github.com/air-supply94/js-algorithm/commit/2bb8e77ad012dc77020bec7495a3609576366c5c))
 - **bst/leetcode:** 迁移到线上 leetcode ([27194fb](https://github.com/air-supply94/js-algorithm/commit/27194fb9690fd9944918c06a207caad3217525be))
 - **bst/leetcode:** 迁移到线上 leetcode ([02bb002](https://github.com/air-supply94/js-algorithm/commit/02bb002c3fc9b2573a3286034f8869c689699925))
@@ -178,6 +179,7 @@
 - **dp/longestCommonSubsequence:** add longestCommonSubsequenceRecursion ([cdf29e3](https://github.com/air-supply94/js-algorithm/commit/cdf29e3f364c3cccdd0632817995c09945e81f44))
 - **dp/longestCommonSubstring:** add longestCommonSubstring ([e3f75fb](https://github.com/air-supply94/js-algorithm/commit/e3f75fbdc9f1ab1ebb0bc141613bdc313f16d119))
 - **dp/longestCreasingSubsequence:** add bs ([41c2be0](https://github.com/air-supply94/js-algorithm/commit/41c2be08eb5f35fdf020e9c51292cf96c35c9100))
+- **dp/maxSubArray:** add getMaxMatrix ([82d31a6](https://github.com/air-supply94/js-algorithm/commit/82d31a69b5445ee0d316d7e1d584a14f03677707))
 - **FenwickTree、binarySearchTreeNode:** add FenwickTree remove binarySearchTreeNode height leftHeight rightHeight banlanceFactory ([7d3763b](https://github.com/air-supply94/js-algorithm/commit/7d3763ba2f08fe52c27f8fe985570bf6e961390d))
 - **findFirstLarge:** add ([90d3278](https://github.com/air-supply94/js-algorithm/commit/90d3278da4215456f0e598654e271dfb2327ae35))
 - **findKthMin:** add ([91879e9](https://github.com/air-supply94/js-algorithm/commit/91879e9a1a5970e46531a1e62833c17efaf6c8e8))